==> data_processing/02_download_energies.ipynb <==
# Download interaction energies

Here we extract interaction energies for the previously located dataset entries:

# imports
import requests
import re, os, time
import pandas as pd
from tqdm import tqdm

# datapaths
path_inp = 'data/entry_IDs.csv'
path_out = 'data/total_energies.csv'

# read entries
if os.path.exists(path_out):
    df = pd.read_csv(path_out)
else:
    df = pd.read_csv('data/entry_IDs.csv')
    df['Total_energy'] = None

df

# URLs
nomad_app = 'https://repository.nomad-coe.eu/app/api'
nomad_rep = f'{nomad_app}/repo'    # calc's metadata
nomad_raw = f'{nomad_app}/raw'     # unprocessed calc's

# params
mainfile = 'FHIaims.out'

def download_energy(upload_id, entry_id):
    '''Returns total QM energy for the entry with given upload and entry IDs'''
    # request entry
    url = f'{nomad_rep}/{upload_id}/{entry_id}'
    metadata = requests.get(url).json()

    # get mainfile
    calc_dir = os.path.dirname(metadata['mainfile'])
    path_mainfile = f'{calc_dir}/{mainfile}'

    # request the mainfile
    url = f'{nomad_raw}/{upload_id}/{path_mainfile}'
    content = requests.get(url).content.decode('utf-8')

    # get energy
    pattern = r'Total energy of the DFT / Hartree-Fock s\.c\.f\. calculation\s*:\s*(-?\d+\.\d+)'
    match = re.search(pattern, content)
    E = float(match.group(1)) if match else None

    return E

# loading params
sub = df.loc[df['Total_energy'].isna()]
n_entries = len(sub)

# loading energies
for i, upload_id, entry_id in tqdm(zip(sub.index, sub['Upload_ID'], sub['Entry_ID']), total = n_entries):
    try:
        E = download_energy(upload_id, entry_id)
    except (KeyboardInterrupt, SystemExit):
        raise
    except:
        time.sleep(30)
        continue
    df.loc[i, 'Total_energy'] = E

# final message
print(f'\nTotal of {sum(df["Total_energy"].isna())} NAs')
print('Download is completed.')

# save data
df.to_csv(path_out, index=False)
df
