==> data_processing/03_processing_energies.ipynb <==
# Processing energies

Here we process extracted data to get interaction energies:

# imports
import pandas as pd

# loading data
df = pd.read_csv('data/total_energies.csv')
df

# get metal and aminoacid
df['Aminoacid'] = df['Mainfile'].apply(lambda x: x.split('/')[1])
df['Cation'] = df['Mainfile'].apply(lambda x: x.split('/')[3])
df = df[['Aminoacid', 'Cation', 'Total_energy']]
df

# get min energies for each system
edf = df.groupby(['Aminoacid', 'Cation']).min().reset_index()
edf

# pivot metals
edf = edf.pivot(index='Aminoacid', columns='Cation', values='Total_energy').reset_index().rename_axis(None, axis = 1)
edf

# combine hisE and hisD
idxs = edf['Aminoacid'].isin(['hisD', 'hisE'])
his = {c: edf.loc[idxs, c].min() for c in edf.columns[1:]}
his['Aminoacid'] = 'his'

# sub hisE and hisD with his
edf = pd.concat([edf.loc[~idxs], pd.DataFrame([his])], ignore_index=True).sort_values('Aminoacid', ignore_index=True)
edf

# compute interaction energies
for M in ('Mg', 'Ca', 'Ba'):
    edf[f'E_{M}'] = edf[M] - edf['bare']
    edf[f'E_{M}'] = edf[f'E_{M}'] - edf.loc[edf['Aminoacid'] == 'gly', f'E_{M}'].values[0]
edf = edf[['Aminoacid', 'E_Mg', 'E_Ca', 'E_Ba']]
edf

# extract ionic systems sidechains
ions = ['lysH', 'argH', 'hisH', 'asp', 'glu']
charged = edf.loc[edf['Aminoacid'].isin(ions)].reset_index(drop=True)
edf = edf.loc[~edf['Aminoacid'].isin(ions)].reset_index(drop=True)
len(edf), len(charged)

# remove 'H' from aminoacid names
for x in (edf, charged):
    x.loc[:,'Aminoacid'] = x['Aminoacid'].str.replace('H', '')
charged = charged.rename(columns = {f'E_{M}': f'E_{M}_ion' for M in ('Mg', 'Ca', 'Ba')})
charged

# add columns for ionized sidechains
for M in ('Mg', 'Ca', 'Ba'):
    edf[f'E_{M}_ion'] = edf[f'E_{M}']

# substitute values for ionised AAs
idxs = edf['Aminoacid'].isin(charged['Aminoacid'])
cols = [f'E_{M}_ion'  for M in ('Mg', 'Ca', 'Ba')]
charged.index = edf.loc[idxs, cols].index
edf.loc[idxs, cols] = charged.loc[:, cols]
edf

# one-letter AA names
AAs = {'ala': 'A', 'arg': 'R', 'asn': 'N', 'asp': 'D',
       'val': 'V', 'his': 'H', 'gly': 'G', 'gln': 'Q',
       'glu': 'E', 'ile': 'I', 'leu': 'L', 'lys': 'K',
       'met': 'M', 'pro': 'P', 'ser': 'S', 'tyr': 'Y',
       'thr': 'T', 'trp': 'W', 'phe': 'F', 'cys': 'C'}
edf['Aminoacid'] = edf['Aminoacid'].replace(AAs)
edf.columns = ['Aminoacid'] + [f'{M} interaction energy, eV' for M in ('Mg', 'Ca', 'Ba')] + \
              [f'{M} interaction energy, charged sidechains, eV' for M in ('Mg', 'Ca', 'Ba')]
edf

# save
edf.to_csv('data/energies.csv', index=False)
