==> data_processing/04_predict_missing_energies.ipynb <==
# Predict interaction energies for missing AAs

Here we use ML to fill the interaction energies for the missing aminoacids:

import pandas as pd
import numpy as np
from rdkit import Chem
from rdkit.Chem import rdMolDescriptors
from rdkit.Chem import Descriptors

from sklearn.ensemble import RandomForestRegressor
from sklearn.impute import SimpleImputer
from sklearn.metrics import mean_absolute_error, r2_score

monomer_smiles = {
    'A': 'CC(C(=O)O)N',  # protein
    'R': 'C(CC(C(=O)O)N)CN=C(N)N',
    'N': 'C(C(C(=O)O)N)C(=O)N',
    'D': 'C(C(C(=O)O)N)C(=O)O',
    'C': 'C(C(C(=O)O)N)S',
    'Q': 'C(CC(=O)N)C(C(=O)O)N',
    'E': 'C(CC(=O)O)C(C(=O)O)N',
    'G': 'C(C(=O)O)N',
    'H': 'C1=C(NC=N1)CC(C(=O)O)N',
    'I': 'CCC(C)C(C(=O)O)N',
    'L': 'CC(C)CC(C(=O)O)N',
    'K': 'C(CCN)CC(C(=O)O)N',
    'M': 'CSCCC(C(=O)O)N',
    'F': 'C1=CC=C(C=C1)CC(C(=O)O)N',
    'P': 'C1CC(NC1)C(=O)O',
    'S': 'C(C(C(=O)O)N)O',
    'T': 'CC(C(C(=O)O)N)O',
    'W': 'C1=CC=C2C(=C1)C(=CN2)CC(C(=O)O)N',
    'Y': 'C1=CC(=CC=C1CC(C(=O)O)N)O',
    'V': 'CC(C)C(C(=O)O)N',
    'O': 'CC1CC=NC1C(=O)NCCCCC(C(=O)O)N',
    'U': 'C(C(C(=O)O)N)[Se]'
}

df = pd.read_csv('data/energies.csv')
df

# add smiles columns
df['smiles'] = df['Aminoacid'].map(monomer_smiles)

# addend for new AAs
additional_aa = [['O', 'CC1CC=NC1C(=O)NCCCCC(C(=O)O)N'],
                 ['U', 'C(C(C(=O)O)N)[Se]']]
df_new = pd.DataFrame(additional_aa, columns=['Aminoacid', 'smiles'])
df_new

data = pd.concat([df, df_new], axis=0).reset_index(drop=True)
data

## Adding rdkit descriptors

To fill missing data we want to use ML algorithms. So we need more descriptors to find distance between data points. Thta's why we are adding new descriptors, later we will drop them.

descriptor_names = list(rdMolDescriptors.Properties.GetAvailableProperties())
get_descriptors = rdMolDescriptors.Properties(descriptor_names)
num_descriptors = len(descriptor_names)

# Initiallization of the empty matrix
descriptors_set = np.empty((0, num_descriptors), float)


for _, row in data.iterrows():
    smiles = row['smiles']
    molecule = Chem.MolFromSmiles(smiles)

    if molecule is not None:
        descriptors = np.array(get_descriptors.ComputeProperties(molecule)).reshape((-1, num_descriptors))
        descriptors_set = np.append(descriptors_set, descriptors, axis=0)

# Creating DataFrame with descriptors
df_descriptors = pd.DataFrame(descriptors_set, columns=descriptor_names)

data = pd.concat([data, df_descriptors], axis=1)

data

# Filling 'Ca interaction energy' column

np.random.seed(1234)

# Define the descriptors and target column
drop_cols = list(data.columns[:8])
descriptors = data.columns.difference(drop_cols)
target_columns = list(data.columns[1:7])

for target_column in target_columns:
    
    # Split the data into training and testing sets
    train_data = data.iloc[:-2]  # Exclude the last two rows with missing values
    test_data = data.iloc[-2:]

    # Create a random forest model to predict missing values
    model = RandomForestRegressor()

    # Train the model on data without missing values in the target column
    model.fit(train_data[descriptors], train_data[target_column])

    # Predict the missing values
    predicted_values = model.predict(test_data[descriptors])

    # Fill in the missing values in the original dataset
    data.loc[data.index[-2:], target_column] = predicted_values

# drop columns
res = data.iloc[:, :7]

# substitute charged sidechains values for U
idxs = res['Aminoacid'] == 'U' # uncharged system
for M in ('Mg', 'Ca', 'Ba'):
    res.loc[idxs, f'{M} interaction energy, charged sidechains, eV'] = res.loc[idxs, f'{M} interaction energy, eV']

res

res.to_csv('data/energies_final.csv', index=False)

# print for README
cols = ['Aminoacid'] + [f'E<sub>int</sub>({M}<sup>2+</sup>), eV' for M in ('Mg', 'Ca', 'Ba')] + \
       [f'E<sub>int</sub>({M}<sup>2+</sup>), charged sidechains, eV' for M in ('Mg', 'Ca', 'Ba')]
print('| ' + ' | '.join(cols) + ' |')
print('| ' + ' | '.join([':---:'] + ['----:']*(len(res.columns) - 1)) + ' |')
for row in res.to_dict('tight')['data']:
    line = '| ' + ' | '.join([row[0]] + [f'{c:>8.3f}' for c in row[1:]]) + ' |'
    print(line)
